Updated all files
Updated README and published latest (unfinished) code files. Test data are images and assorted ROIs used to test the functions.

ROITools.py is the main file where I am building out the functions. It is a work in progress.

ROI_GUI.py is a start at creating a GUI for this program to speed up analysis of data.